add escaping for file paths
